Changed `expression::Builder::build()` back to returning an `Expression` rather than a `Result`
I decided it's probably easier for the user to look at the resulting error from DynamoDB to determine where the empty path is.

Implementing a more complicated solution to return an `Err` to give the user more clarity about what path is empty would make the ergonomics of the crate too cumbersome.

This is a place we will allow the user to do the wrong thing.